+++ src/frontend/apps/impress/src/features/docs/doc-management/api/index.ts
@@ -1,3 +1,4 @@
+export * from './useCreateDoc';
 export * from './useDoc';
 export * from './useDocs';
 export * from './useUpdateDoc';