+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 
 - ✨(ci) add security scan #291
 - ✨(frontend) Activate versions feature #240
+- ✨(frontend) one-click document creation #275
 
 ## Changed